Feature: receive_string millis overflow tested
Receive_string will now timeout correctly when millis overflows
	modified:   src/uart_2_controller.c
	modified:   src/uart_2_controller.h
	modified:   test/test_uart_2_controller.c

+++ src/uart_2_controller.h
@@ -8,7 +8,7 @@
 /**
  * Sets the timeout period for expecting a string to occur
  */
-void uart_2_controller_set_timeout(uint16_t timeout);
+void uart_2_controller_set_timeout(uint32_t timeout);
 
 void uart_2_controller_send_string(char * string, uint8_t length);
